Create CMakeLists and Readme file, fix size_t warnings

+++ hnetw/capture/hn_packet_capturer.cpp
@@ -80,7 +80,7 @@ void HnPacketCapturer::capturePackets()
 {
     if (!socketSetToCapture_) return;
 
-    int bytesRead = 0;
+    size_t bytesRead = 0;
     std::clock_t currentPacketTime = 0;
 
     char* buffer = nullptr;
